test(ng-dev): refactor the testTmpDir to be defined in utils/testing
Moves all of the references and instances of testTmpDir to be centralized to using one common symbol.
Previously it was defined in two different places and ended up being two different directories.
