+++ ng-dev/release/publish/test/common.spec.ts
@@ -11,7 +11,7 @@ import {join} from 'path';
 import * as semver from 'semver';
 import {SemVer} from 'semver';
 
-import {getBranchPushMatcher} from '../../../utils/testing';
+import {getBranchPushMatcher, testTmpDir} from '../../../utils/testing';
 import {ReleaseNotes} from '../../notes/release-notes';
 import {NpmDistTag} from '../../versioning';
 import {ActiveReleaseTrains} from '../../versioning/active-release-trains';
@@ -26,15 +26,11 @@ import {
   parse,
   setupReleaseActionForTesting,
 } from './test-utils/test-utils';
-import {
-  getTestConfigurationsForAction,
-  testReleasePackages,
-  testTmpDir,
-} from './test-utils/action-mocks';
+import {getTestConfigurationsForAction, testReleasePackages} from './test-utils/action-mocks';
 import {getMockGitClient} from './test-utils/git-client-mock';
 import {CommitFromGitLog, parseCommitFromGitLog} from '../../../commit-message/parse';
 import {SandboxGitRepo} from './test-utils/sandbox-testing';
-import { GitClient } from '../../../utils/git/git-client';
+import {GitClient} from '../../../utils/git/git-client';
 
 describe('common release action logic', () => {
   const baseReleaseTrains: ActiveReleaseTrains = {
@@ -143,7 +139,10 @@ describe('common release action logic', () => {
     });
 
     it('should link to the changelog in the release entry if notes are too large', async () => {
-      const {repo, instance, gitClient} = setupReleaseActionForTesting(TestAction, baseReleaseTrains);
+      const {repo, instance, gitClient} = setupReleaseActionForTesting(
+        TestAction,
+        baseReleaseTrains,
+      );
       const {version, branchName} = baseReleaseTrains.latest;
       const tagName = version.format();
       const testCommit = parseCommitFromGitLog(Buffer.from('fix(test): test'));
@@ -158,7 +157,7 @@ describe('common release action logic', () => {
       testCommit.subject = exceedingText;
 
       spyOn(ReleaseNotes, 'forRange').and.callFake(
-        async () => new MockReleaseNotes(version, [testCommit], gitClient)
+        async () => new MockReleaseNotes(version, [testCommit], gitClient),
       );
 
       repo
