+++ ng-dev/release/publish/test/cut-lts-patch.spec.ts
@@ -20,10 +20,11 @@ import {
   expectGithubApiRequestsForStaging,
   expectStagingAndPublishWithCherryPick,
 } from './test-utils/staging-test';
-import {getTestConfigurationsForAction, testTmpDir} from './test-utils/action-mocks';
+import {getTestConfigurationsForAction} from './test-utils/action-mocks';
 import {SandboxGitRepo} from './test-utils/sandbox-testing';
 import {readFileSync} from 'fs';
 import {getMockGitClient} from './test-utils/git-client-mock';
+import {testTmpDir} from '../../../utils/testing';
 
 describe('cut an LTS patch action', () => {
   it('should be active', async () => {
