+++ ng-dev/release/publish/test/cut-new-patch.spec.ts
@@ -15,7 +15,7 @@ import {
 } from './test-utils/staging-test';
 import {SandboxGitRepo} from './test-utils/sandbox-testing';
 import {readFileSync} from 'fs';
-import {testTmpDir} from './test-utils/action-mocks';
+import {testTmpDir} from '../../../utils/testing';
 
 describe('cut new patch action', () => {
   it('should be active', async () => {
