+++ ng-dev/release/publish/test/cut-release-candidate-for-feature-freeze.spec.ts
@@ -14,8 +14,8 @@ import {
   expectStagingAndPublishWithCherryPick,
 } from './test-utils/staging-test';
 import {readFileSync} from 'fs';
-import {testTmpDir} from './test-utils/action-mocks';
 import {SandboxGitRepo} from './test-utils/sandbox-testing';
+import {testTmpDir} from '../../../utils/testing';
 
 describe('cut release candidate for feature-freeze action', () => {
   it('should activate if a feature-freeze release-train is active', async () => {
