+++ ng-dev/release/publish/test/test-utils/test-utils.ts
@@ -9,16 +9,13 @@
 import * as nock from 'nock';
 import * as semver from 'semver';
 
-import {
-  getTestConfigurationsForAction,
-  setupMocksForReleaseAction,
-  testTmpDir,
-} from './action-mocks';
+import {getTestConfigurationsForAction, setupMocksForReleaseAction} from './action-mocks';
 import {_npmPackageInfoCache, ActiveReleaseTrains, NpmPackageInfo} from '../../../versioning';
 import {ReleaseAction, ReleaseActionConstructor} from '../../actions';
 import {GithubTestingRepo} from './github-api-testing';
 import {defaultTestOptions, TestOptions, TestReleaseAction} from './test-action';
 import {dedent} from '../../../../utils/testing/dedent';
+import {testTmpDir} from '../../../../utils/testing';
 
 /**
  * Sets up the given release action for testing.
